+++ src/app/account/page.tsx
@@ -117,7 +117,7 @@ export default function Account() {
               <>
                 {!PWAReady && !isPWA && (
                   <Typography color="danger" textAlign="center">
-                    PWAを起動中です。しばらくしてから再度お試しください。
+                    PWAを準備中です。しばらくしてからページを更新してください。
                   </Typography>
                 )}
                 {(user?.loginType === "Guest" || !user?.isMailVerified) && (
